Start of refactor to stash implementation to match other remotes
