+++ remote/bitbucketserver/bitbucketserver.go
@@ -12,13 +12,23 @@ import (
 	"io/ioutil"
 	"net/http"
 	"net/url"
-
-	log "github.com/Sirupsen/logrus"
+ 	"github.com/drone/drone/remote/bitbucketserver/internal"
 	"github.com/drone/drone/model"
 	"github.com/drone/drone/remote"
 	"github.com/mrjones/oauth"
 	"strings"
+	"crypto/tls"
+	"encoding/hex"
+	"crypto/md5"
 )
+
+const (
+	requestTokenURL = "%s/plugins/servlet/oauth/request-token"
+	authorizeTokenURL = "%s/plugins/servlet/oauth/authorize"
+	accessTokenURL = "%s/plugins/servlet/oauth/access-token"
+)
+
+
 
 // Opts defines configuration options.
 type Opts struct {
@@ -27,218 +37,121 @@ type Opts struct {
 	Password    string // Git machine account password.
 	ConsumerKey string // Oauth1 consumer key.
 	ConsumerRSA string // Oauth1 consumer key file.
-
 	SkipVerify bool // Skip ssl verification.
+}
+
+type Config struct {
+	URL string
+	Username string
+	Password string
+	PrivateKey *rsa.PrivateKey
+	ConsumerKey string
+	SkipVerify bool
+
 }
 
 // New returns a Remote implementation that integrates with Bitbucket Server,
 // the on-premise edition of Bitbucket Cloud, formerly known as Stash.
 func New(opts Opts) (remote.Remote, error) {
-	bb := &client{
-		URL:         opts.URL,
+	config := &Config{
+		URL: opts.URL,
+		Username: opts.Username,
+		Password: opts.Password,
 		ConsumerKey: opts.ConsumerKey,
-		ConsumerRSA: opts.ConsumerRSA,
-		GitUserName: opts.Username,
-		GitPassword: opts.Password,
+		SkipVerify: opts.SkipVerify,
 	}
 
 	switch {
-	case bb.GitUserName == "":
+	case opts.Username == "":
 		return nil, fmt.Errorf("Must have a git machine account username")
-	case bb.GitPassword == "":
+	case opts.Password == "":
 		return nil, fmt.Errorf("Must have a git machine account password")
-	case bb.ConsumerKey == "":
+	case opts.ConsumerKey == "":
 		return nil, fmt.Errorf("Must have a oauth1 consumer key")
-	case bb.ConsumerRSA == "":
+	case opts.ConsumerRSA == "":
 		return nil, fmt.Errorf("Must have a oauth1 consumer key file")
 	}
 
-	keyfile, err := ioutil.ReadFile(bb.ConsumerRSA)
+	keyFile, err := ioutil.ReadFile(opts.ConsumerRSA)
 	if err != nil {
 		return nil, err
 	}
-	block, _ := pem.Decode(keyfile)
-	bb.PrivateKey, err = x509.ParsePKCS1PrivateKey(block.Bytes)
+	block, _ := pem.Decode(keyFile)
+	config.PrivateKey, err = x509.ParsePKCS1PrivateKey(block.Bytes)
 	if err != nil {
 		return nil, err
 	}
-
-	// TODO de-referencing is a bit weird and may not behave as expected, and could
-	// have race conditions. Instead store the parsed key (I already did this above)
-	// and then pass the parsed private key when creating the Bitbucket client.
-	bb.Consumer = *NewClient(bb.ConsumerRSA, bb.ConsumerKey, bb.URL)
-	return bb, nil
+	return config, nil
 }
 
-type client struct {
-	URL         string
-	ConsumerKey string
-	GitUserName string
-	GitPassword string
-	ConsumerRSA string
-	PrivateKey  *rsa.PrivateKey
-	Consumer    oauth.Consumer
-}
 
-func (c *client) Login(res http.ResponseWriter, req *http.Request) (*model.User, error) {
-	requestToken, url, err := c.Consumer.GetRequestTokenAndUrl("oob")
+func (c *Config) Login(res http.ResponseWriter, req *http.Request) (*model.User, error) {
+	requestToken, url, err := c.Consumer().GetRequestTokenAndUrl("oob")
 	if err != nil {
 		return nil, err
 	}
-
 	var code = req.FormValue("oauth_verifier")
 	if len(code) == 0 {
 		http.Redirect(res, req, url, http.StatusSeeOther)
 		return nil, nil
 	}
-
 	requestToken.Token = req.FormValue("oauth_token")
-	accessToken, err := c.Consumer.AuthorizeToken(requestToken, code)
+	accessToken, err := c.Consumer().AuthorizeToken(requestToken, code)
 	if err != nil {
 		return nil, err
 	}
 
-	client, err := c.Consumer.MakeHttpClient(accessToken)
-	if err != nil {
-		return nil, err
-	}
+	client := internal.NewClientWithToken(c.URL, c.Consumer(), accessToken.Token)
 
-	response, err := client.Get(fmt.Sprintf("%s/plugins/servlet/applinks/whoami", c.URL))
-	if err != nil {
-		return nil, err
-	}
-	defer response.Body.Close()
-	bits, err := ioutil.ReadAll(response.Body)
-	if err != nil {
-		return nil, err
-	}
-	login := string(bits)
+	return client.FindCurrentUser()
 
-	// TODO errors should never be ignored like this
-	response1, err := client.Get(fmt.Sprintf("%s/rest/api/1.0/users/%s", c.URL, login))
-	if err != nil {
-		return nil, err
-	}
-	defer response1.Body.Close()
 
-	contents, err := ioutil.ReadAll(response1.Body)
-	if err !=nil {
-		return nil, err
-	}
-
-	var user User
-	err = json.Unmarshal(contents, &user)
-	if err != nil {
-		return nil, err
-	}
-	return &model.User{
-		Login:  login,
-		Email:  user.EmailAddress,
-		Token:  accessToken.Token,
-		Avatar: avatarLink(user.EmailAddress),
-	}, nil
 }
 
 // Auth is not supported by the Stash driver.
-func (*client) Auth(token, secret string) (string, error) {
+func (*Config) Auth(token, secret string) (string, error) {
 	return "", fmt.Errorf("Not Implemented")
 }
 
 // Teams is not supported by the Stash driver.
-func (*client) Teams(u *model.User) ([]*model.Team, error) {
+func (*Config) Teams(u *model.User) ([]*model.Team, error) {
 	var teams []*model.Team
 	return teams, nil
 }
 
-func (c *client) Repo(u *model.User, owner, name string) (*model.Repo, error) {
-	client := NewClientWithToken(&c.Consumer, u.Token)
-	repo , err := c.FindRepo(client,owner,name)
-	if err != nil {
-		return nil, err
-	}
-	return repo, nil
+func (c *Config) Repo(u *model.User, owner, name string) (*model.Repo, error) {
+
+	client := internal.NewClientWithToken(c.URL, c.Consumer(), u.Token)
+
+	return client.FindRepo(owner, name)
 }
 
-func (c *client) Repos(u *model.User) ([]*model.RepoLite, error) {
+func (c *Config) Repos(u *model.User) ([]*model.RepoLite, error) {
 
-	var repos = []*model.RepoLite{}
+	client := internal.NewClientWithToken(c.URL,c.Consumer(), u.Token)
 
-	client := NewClientWithToken(&c.Consumer, u.Token)
-
-	response, err := client.Get(fmt.Sprintf("%s/rest/api/1.0/repos?limit=10000", c.URL))
-	if err != nil {
-		log.Error(err)
-	}
-	defer response.Body.Close()
-	contents, err := ioutil.ReadAll(response.Body)
-	if err != nil {
-		return nil, err
-	}
-	var repoResponse Repos
-	err = json.Unmarshal(contents, &repoResponse)
-	if err != nil {
-		return nil, err
-	}
-
-	for _, repo := range repoResponse.Values {
-		repos = append(repos, &model.RepoLite{
-			Name:     repo.Slug,
-			FullName: repo.Project.Key + "/" + repo.Slug,
-			Owner:    repo.Project.Key,
-		})
-	}
-
-	return repos, nil
+	return client.FindRepos()
 }
 
-func (c *client) Perm(u *model.User, owner, repo string) (*model.Perm, error) {
-	client := NewClientWithToken(&c.Consumer, u.Token)
-	perms := new(model.Perm)
+func (c *Config) Perm(u *model.User, owner, repo string) (*model.Perm, error) {
+	client := internal.NewClientWithToken(c.URL,c.Consumer(), u.Token)
 
-	// If you don't have access return none right away
-	_, err := c.FindRepo(client, owner, repo)
-	if err != nil {
-		return perms, err
-	}
-
-	// Must have admin to be able to list hooks. If have access the enable perms
-	_, err = client.Get(fmt.Sprintf("%s/rest/api/1.0/projects/%s/repos/%s/settings/hooks/%s", c.URL, owner, repo,"com.atlassian.stash.plugin.stash-web-post-receive-hooks-plugin:postReceiveHook"))
-	if err == nil {
-		perms.Push = true
-		perms.Admin = true
-	}
-	perms.Pull = true
-	return perms, nil
+	return client.FindRepoPerms(owner, repo)
 }
 
-func (c *client) File(u *model.User, r *model.Repo, b *model.Build, f string) ([]byte, error) {
-	log.Info(fmt.Sprintf("Staring file for bitbucketServer login: %s repo: %s buildevent: %s string: %s", u.Login, r.Name, b.Event, f))
+func (c *Config) File(u *model.User, r *model.Repo, b *model.Build, f string) ([]byte, error) {
 
-	client := NewClientWithToken(&c.Consumer, u.Token)
-	fileURL := fmt.Sprintf("%s/projects/%s/repos/%s/browse/%s?raw", c.URL, r.Owner, r.Name, f)
-	log.Info(fileURL)
-	response, err := client.Get(fileURL)
-	if err != nil {
-		log.Error(err)
-	}
-	if response.StatusCode == 404 {
-		return nil, nil
-	}
-	defer response.Body.Close()
-	responseBytes, err := ioutil.ReadAll(response.Body)
-	if err != nil {
-		log.Error(err)
-	}
+	client := internal.NewClientWithToken(c.URL, c.Consumer(), u.Token)
 
-	return responseBytes, nil
+	return client.FindFileForRepo(r.Owner, r.Name, f)
 }
 
 // Status is not supported by the Gogs driver.
-func (*client) Status(*model.User, *model.Repo, *model.Build, string) error {
+func (*Config) Status(*model.User, *model.Repo, *model.Build, string) error {
 	return nil
 }
 
-func (c *client) Netrc(user *model.User, r *model.Repo) (*model.Netrc, error) {
+func (c *Config) Netrc(user *model.User, r *model.Repo) (*model.Netrc, error) {
 	u, err := url.Parse(c.URL)
 	if err != nil {
 		return nil, err
@@ -252,31 +165,23 @@ func (c *client) Netrc(user *model.User, r *model.Repo) (*model.Netrc, error) {
 	}
 	return &model.Netrc{
 		Machine:  host,
-		Login:    c.GitUserName,
-		Password: c.GitPassword,
+		Login:    c.Username,
+		Password: c.Password,
 	}, nil
 }
 
-func (c *client) Activate(u *model.User, r *model.Repo, link string) error {
-	client := NewClientWithToken(&c.Consumer, u.Token)
-	hook, err := c.CreateHook(client, r.Owner, r.Name, "com.atlassian.stash.plugin.stash-web-post-receive-hooks-plugin:postReceiveHook", link)
-	if err != nil {
-		return err
-	}
-	log.Info(hook)
-	return nil
+func (c *Config) Activate(u *model.User, r *model.Repo, link string) error {
+	client := internal.NewClientWithToken(c.URL, c.Consumer(), u.Token)
+
+	return client.CreateHook(r.Owner, r.Name, link)
 }
 
-func (c *client) Deactivate(u *model.User, r *model.Repo, link string) error {
-	client := NewClientWithToken(&c.Consumer, u.Token)
-	err := c.DeleteHook(client, r.Owner, r.Name, "com.atlassian.stash.plugin.stash-web-post-receive-hooks-plugin:postReceiveHook", link)
-	if err != nil {
-		return err
-	}
-	return nil
+func (c *Config) Deactivate(u *model.User, r *model.Repo, link string) error {
+	client := internal.NewClientWithToken(c.URL, c.Consumer(), u.Token)
+	return client.DeleteHook(r.Owner, r.Name, link)
 }
 
-func (c *client) Hook(r *http.Request) (*model.Repo, *model.Build, error) {
+func (c *Config) Hook(r *http.Request) (*model.Repo, *model.Build, error) {
 	hook := new(postHook)
 	if err := json.NewDecoder(r.Body).Decode(hook); err != nil {
 		return nil, nil, err
@@ -301,83 +206,29 @@ func (c *client) Hook(r *http.Request) (*model.Repo, *model.Build, error) {
 	return repo, build, nil
 }
 
-type HookDetail struct {
-	Key           string `json:"key"`
-	Name          string `json:"name"`
-	Type          string `json:"type"`
-	Description   string `json:"description"`
-	Version       string `json:"version"`
-	ConfigFormKey string `json:"configFormKey"`
+
+func (c *Config) Consumer()  *oauth.Consumer{
+	consumer := oauth.NewRSAConsumer(
+		c.ConsumerKey,
+		c.PrivateKey,
+		oauth.ServiceProvider{
+			RequestTokenUrl:   fmt.Sprintf(requestTokenURL, c.URL),
+			AuthorizeTokenUrl: fmt.Sprintf(authorizeTokenURL, c.URL),
+			AccessTokenUrl:    fmt.Sprintf(accessTokenURL, c.URL),
+			HttpMethod:        "POST",
+		})
+	consumer.HttpClient = &http.Client{
+		Transport: &http.Transport{
+			TLSClientConfig: &tls.Config{InsecureSkipVerify: true},
+		},
+	}
+	return consumer
 }
 
-type Hook struct {
-	Enabled bool        `json:"enabled"`
-	Details *HookDetail `json:"details"`
-}
-
-// Enable hook for named repository
-func (bs *client) CreateHook(client *http.Client, project, slug, hook_key, link string) (*Hook, error) {
-
-	// Set hook
-	hookBytes := []byte(fmt.Sprintf(`{"hook-url-0":"%s"}`, link))
-
-	// Enable hook
-	enablePath := fmt.Sprintf("/rest/api/1.0/projects/%s/repos/%s/settings/hooks/%s/enabled",
-		project, slug, hook_key)
-
-	doPut(client, bs.URL+enablePath, hookBytes)
-
-	return nil, nil
-}
-
-// Disable hook for named repository
-func (bs *client) DeleteHook(client *http.Client, project, slug, hook_key, link string) error {
-	enablePath := fmt.Sprintf("/rest/api/1.0/projects/%s/repos/%s/settings/hooks/%s/enabled",
-		project, slug, hook_key)
-	doDelete(client, bs.URL+enablePath)
-
-	return nil
-}
-
-func (c *client) FindRepo(client *http.Client, owner string, name string) (*model.Repo, error){
-
-	urlString := fmt.Sprintf("%s/rest/api/1.0/projects/%s/repos/%s", c.URL, owner, name)
-
-	response, err := client.Get(urlString)
-	if err != nil {
-		log.Error(err)
-	}
-	defer response.Body.Close()
-	contents, err := ioutil.ReadAll(response.Body)
-	bsRepo := BSRepo{}
-	err = json.Unmarshal(contents, &bsRepo)
-	if err !=nil {
-		return nil, err
-	}
-	repo := &model.Repo{
-		Name:      bsRepo.Slug,
-		Owner:     bsRepo.Project.Key,
-		Branch:    "master",
-		Kind:      model.RepoGit,
-		IsPrivate: true, // TODO(josmo) possibly set this as a setting - must always be private to use netrc
-		FullName:  fmt.Sprintf("%s/%s", bsRepo.Project.Key, bsRepo.Slug),
-	}
-
-	for _, item := range bsRepo.Links.Clone {
-		if item.Name == "http" {
-			uri, err := url.Parse(item.Href)
-			if err != nil {
-				return nil, err
-			}
-			uri.User = nil
-			repo.Clone = uri.String()
-		}
-	}
-	for _, item := range bsRepo.Links.Self {
-		if item.Href != "" {
-			repo.Link = item.Href
-		}
-	}
-
-	return repo, nil
+func avatarLink(email string) (url string) {
+	hasher := md5.New()
+	hasher.Write([]byte(strings.ToLower(email)))
+	emailHash := fmt.Sprintf("%v", hex.EncodeToString(hasher.Sum(nil)))
+	avatarURL := fmt.Sprintf("https://www.gravatar.com/avatar/%s.jpg", emailHash)
+	return avatarURL
 }
